+++ sdk/tables/Azure.Data.Tables/src/TableConstants.cs
@@ -10,6 +10,7 @@ namespace Azure.Data.Tables
             public const string Date = "x-ms-date";
             public const string SharedKey = "SharedKeyLite";
             public const string Authorization = "Authorization";
+            public const string IfMatch = "If-Match";
         }
 
         internal static class QueryParameterNames
@@ -17,6 +18,14 @@ namespace Azure.Data.Tables
             public const string NextTableName = "NextTableName";
             public const string NextPartitionKey = "NextPartitionKey";
             public const string NextRowKey = "NextRowKey";
+        }
+
+        internal static class PropertyNames
+        {
+            public const string TimeStamp = "Timestamp";
+            public const string PartitionKey = "PartitionKey";
+            public const string RowKey = "RowKey";
+            public const string Etag = "odata.etag";
         }
     }
 }